requirements.txt
asgiref==3.7.2
Django==5.0.2
django-rest-framework==0.1.0
djangorestframework==3.15.2
pillow==10.2.0
PyJWT==2.9.0
sqlparse==0.4.4
typing_extensions==4.9.0
